driver: move api to docker pkg
