+++ main.go
@@ -7,7 +7,6 @@ import (
 	"github.com/docker/go-plugins-helpers/sdk"
 	"github.com/sirupsen/logrus"
 
-	"github.com/rchicoli/docker-log-elasticsearch/api"
 	"github.com/rchicoli/docker-log-elasticsearch/pkg/docker"
 )
 
@@ -32,7 +31,7 @@ func main() {
 
 	h := sdk.NewHandler(`{"Implements": ["LoggingDriver"]}`)
 	d := docker.NewDriver()
-	api.Handlers(&h, d)
+	docker.Handlers(&h, d)
 	if err := h.ServeUnix(d.Name(), 0); err != nil {
 		panic(err)
 	}
